+++ CMakeLists.txt
@@ -564,6 +564,7 @@ set(UNICORN_ARCH_COMMON
     qemu/accel/tcg/tcg-all.c
     qemu/accel/tcg/tcg-runtime.c
     qemu/accel/tcg/tcg-runtime-gvec.c
+    qemu/accel/tcg/tb-maint.c
     qemu/accel/tcg/translate-all.c
     qemu/accel/tcg/translator.c
