+++ qemu/aarch64.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_aarch64
 #define tlb_reset_dirty tlb_reset_dirty_aarch64
 #define tlb_set_dirty tlb_set_dirty_aarch64
+#define tlb_set_page_full tlb_set_page_full_aarch64
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_aarch64
 #define tlb_set_page tlb_set_page_aarch64
 #define get_page_addr_code_hostp get_page_addr_code_hostp_aarch64
 #define get_page_addr_code get_page_addr_code_aarch64
 #define probe_access probe_access_aarch64
+#define probe_access_flags probe_access_flags_aarch64
+#define probe_access_full probe_access_full_aarch64
 #define tlb_vaddr_to_host tlb_vaddr_to_host_aarch64
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_aarch64
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_aarch64
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_aarch64
 #define helper_gvec_bitsel helper_gvec_bitsel_aarch64
 #define cpu_restore_state cpu_restore_state_aarch64
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_aarch64
+#define page_find_alloc page_find_alloc_aarch64
+#define page_lock page_lock_aarch64
+#define page_unlock page_unlock_aarch64
+#define do_assert_page_locked do_assert_page_locked_aarch64
+#define assert_no_pages_locked assert_no_pages_locked_aarch64
 #define page_collection_lock page_collection_lock_aarch64
 #define page_collection_unlock page_collection_unlock_aarch64
 #define free_code_gen_buffer free_code_gen_buffer_aarch64
 #define tcg_exec_init tcg_exec_init_aarch64
+#define tb_htable_init tb_htable_init_aarch64
 #define tb_cleanup tb_cleanup_aarch64
 #define tb_flush tb_flush_aarch64
 #define tb_phys_invalidate tb_phys_invalidate_aarch64
 #define tb_gen_code tb_gen_code_aarch64
+#define tb_reset_jump tb_reset_jump_aarch64
+#define tb_link_page tb_link_page_aarch64
 #define tb_exec_lock tb_exec_lock_aarch64
 #define tb_exec_unlock tb_exec_unlock_aarch64
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_aarch64
