+++ qemu/arm.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_arm
 #define tlb_reset_dirty tlb_reset_dirty_arm
 #define tlb_set_dirty tlb_set_dirty_arm
+#define tlb_set_page_full tlb_set_page_full_arm
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_arm
 #define tlb_set_page tlb_set_page_arm
 #define get_page_addr_code_hostp get_page_addr_code_hostp_arm
 #define get_page_addr_code get_page_addr_code_arm
 #define probe_access probe_access_arm
+#define probe_access_flags probe_access_flags_arm
+#define probe_access_full probe_access_full_arm
 #define tlb_vaddr_to_host tlb_vaddr_to_host_arm
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_arm
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_arm
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_arm
 #define helper_gvec_bitsel helper_gvec_bitsel_arm
 #define cpu_restore_state cpu_restore_state_arm
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_arm
+#define page_find_alloc page_find_alloc_arm
+#define page_lock page_lock_arm
+#define page_unlock page_unlock_arm
+#define do_assert_page_locked do_assert_page_locked_arm
+#define assert_no_pages_locked assert_no_pages_locked_arm
 #define page_collection_lock page_collection_lock_arm
 #define page_collection_unlock page_collection_unlock_arm
 #define free_code_gen_buffer free_code_gen_buffer_arm
 #define tcg_exec_init tcg_exec_init_arm
+#define tb_htable_init tb_htable_init_arm
 #define tb_cleanup tb_cleanup_arm
 #define tb_flush tb_flush_arm
 #define tb_phys_invalidate tb_phys_invalidate_arm
 #define tb_gen_code tb_gen_code_arm
+#define tb_reset_jump tb_reset_jump_arm
+#define tb_link_page tb_link_page_arm
 #define tb_exec_lock tb_exec_lock_arm
 #define tb_exec_unlock tb_exec_unlock_arm
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_arm
