+++ qemu/m68k.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_m68k
 #define tlb_reset_dirty tlb_reset_dirty_m68k
 #define tlb_set_dirty tlb_set_dirty_m68k
+#define tlb_set_page_full tlb_set_page_full_m68k
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_m68k
 #define tlb_set_page tlb_set_page_m68k
 #define get_page_addr_code_hostp get_page_addr_code_hostp_m68k
 #define get_page_addr_code get_page_addr_code_m68k
 #define probe_access probe_access_m68k
+#define probe_access_flags probe_access_flags_m68k
+#define probe_access_full probe_access_full_m68k
 #define tlb_vaddr_to_host tlb_vaddr_to_host_m68k
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_m68k
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_m68k
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_m68k
 #define helper_gvec_bitsel helper_gvec_bitsel_m68k
 #define cpu_restore_state cpu_restore_state_m68k
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_m68k
+#define page_find_alloc page_find_alloc_m68k
+#define page_lock page_lock_m68k
+#define page_unlock page_unlock_m68k
+#define do_assert_page_locked do_assert_page_locked_m68k
+#define assert_no_pages_locked assert_no_pages_locked_m68k
 #define page_collection_lock page_collection_lock_m68k
 #define page_collection_unlock page_collection_unlock_m68k
 #define free_code_gen_buffer free_code_gen_buffer_m68k
 #define tcg_exec_init tcg_exec_init_m68k
+#define tb_htable_init tb_htable_init_m68k
 #define tb_cleanup tb_cleanup_m68k
 #define tb_flush tb_flush_m68k
 #define tb_phys_invalidate tb_phys_invalidate_m68k
 #define tb_gen_code tb_gen_code_m68k
+#define tb_reset_jump tb_reset_jump_m68k
+#define tb_link_page tb_link_page_m68k
 #define tb_exec_lock tb_exec_lock_m68k
 #define tb_exec_unlock tb_exec_unlock_m68k
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_m68k
