+++ qemu/mips64el.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_mips64el
 #define tlb_reset_dirty tlb_reset_dirty_mips64el
 #define tlb_set_dirty tlb_set_dirty_mips64el
+#define tlb_set_page_full tlb_set_page_full_mips64el
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_mips64el
 #define tlb_set_page tlb_set_page_mips64el
 #define get_page_addr_code_hostp get_page_addr_code_hostp_mips64el
 #define get_page_addr_code get_page_addr_code_mips64el
 #define probe_access probe_access_mips64el
+#define probe_access_flags probe_access_flags_mips64el
+#define probe_access_full probe_access_full_mips64el
 #define tlb_vaddr_to_host tlb_vaddr_to_host_mips64el
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_mips64el
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_mips64el
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_mips64el
 #define helper_gvec_bitsel helper_gvec_bitsel_mips64el
 #define cpu_restore_state cpu_restore_state_mips64el
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_mips64el
+#define page_find_alloc page_find_alloc_mips64el
+#define page_lock page_lock_mips64el
+#define page_unlock page_unlock_mips64el
+#define do_assert_page_locked do_assert_page_locked_mips64el
+#define assert_no_pages_locked assert_no_pages_locked_mips64el
 #define page_collection_lock page_collection_lock_mips64el
 #define page_collection_unlock page_collection_unlock_mips64el
 #define free_code_gen_buffer free_code_gen_buffer_mips64el
 #define tcg_exec_init tcg_exec_init_mips64el
+#define tb_htable_init tb_htable_init_mips64el
 #define tb_cleanup tb_cleanup_mips64el
 #define tb_flush tb_flush_mips64el
 #define tb_phys_invalidate tb_phys_invalidate_mips64el
 #define tb_gen_code tb_gen_code_mips64el
+#define tb_reset_jump tb_reset_jump_mips64el
+#define tb_link_page tb_link_page_mips64el
 #define tb_exec_lock tb_exec_lock_mips64el
 #define tb_exec_unlock tb_exec_unlock_mips64el
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_mips64el
