+++ qemu/s390x.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_s390x
 #define tlb_reset_dirty tlb_reset_dirty_s390x
 #define tlb_set_dirty tlb_set_dirty_s390x
+#define tlb_set_page_full tlb_set_page_full_s390x
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_s390x
 #define tlb_set_page tlb_set_page_s390x
 #define get_page_addr_code_hostp get_page_addr_code_hostp_s390x
 #define get_page_addr_code get_page_addr_code_s390x
 #define probe_access probe_access_s390x
+#define probe_access_flags probe_access_flags_s390x
+#define probe_access_full probe_access_full_s390x
 #define tlb_vaddr_to_host tlb_vaddr_to_host_s390x
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_s390x
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_s390x
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_s390x
 #define helper_gvec_bitsel helper_gvec_bitsel_s390x
 #define cpu_restore_state cpu_restore_state_s390x
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_s390x
+#define page_find_alloc page_find_alloc_s390x
+#define page_lock page_lock_s390x
+#define page_unlock page_unlock_s390x
+#define do_assert_page_locked do_assert_page_locked_s390x
+#define assert_no_pages_locked assert_no_pages_locked_s390x
 #define page_collection_lock page_collection_lock_s390x
 #define page_collection_unlock page_collection_unlock_s390x
 #define free_code_gen_buffer free_code_gen_buffer_s390x
 #define tcg_exec_init tcg_exec_init_s390x
+#define tb_htable_init tb_htable_init_s390x
 #define tb_cleanup tb_cleanup_s390x
 #define tb_flush tb_flush_s390x
 #define tb_phys_invalidate tb_phys_invalidate_s390x
 #define tb_gen_code tb_gen_code_s390x
+#define tb_reset_jump tb_reset_jump_s390x
+#define tb_link_page tb_link_page_s390x
 #define tb_exec_lock tb_exec_lock_s390x
 #define tb_exec_unlock tb_exec_unlock_s390x
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_s390x
