+++ qemu/sparc.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_sparc
 #define tlb_reset_dirty tlb_reset_dirty_sparc
 #define tlb_set_dirty tlb_set_dirty_sparc
+#define tlb_set_page_full tlb_set_page_full_sparc
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_sparc
 #define tlb_set_page tlb_set_page_sparc
 #define get_page_addr_code_hostp get_page_addr_code_hostp_sparc
 #define get_page_addr_code get_page_addr_code_sparc
 #define probe_access probe_access_sparc
+#define probe_access_flags probe_access_flags_sparc
+#define probe_access_full probe_access_full_sparc
 #define tlb_vaddr_to_host tlb_vaddr_to_host_sparc
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_sparc
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_sparc
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_sparc
 #define helper_gvec_bitsel helper_gvec_bitsel_sparc
 #define cpu_restore_state cpu_restore_state_sparc
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_sparc
+#define page_find_alloc page_find_alloc_sparc
+#define page_lock page_lock_sparc
+#define page_unlock page_unlock_sparc
+#define do_assert_page_locked do_assert_page_locked_sparc
+#define assert_no_pages_locked assert_no_pages_locked_sparc
 #define page_collection_lock page_collection_lock_sparc
 #define page_collection_unlock page_collection_unlock_sparc
 #define free_code_gen_buffer free_code_gen_buffer_sparc
 #define tcg_exec_init tcg_exec_init_sparc
+#define tb_htable_init tb_htable_init_sparc
 #define tb_cleanup tb_cleanup_sparc
 #define tb_flush tb_flush_sparc
 #define tb_phys_invalidate tb_phys_invalidate_sparc
 #define tb_gen_code tb_gen_code_sparc
+#define tb_reset_jump tb_reset_jump_sparc
+#define tb_link_page tb_link_page_sparc
 #define tb_exec_lock tb_exec_lock_sparc
 #define tb_exec_unlock tb_exec_unlock_sparc
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_sparc
