+++ qemu/tricore.h
@@ -883,11 +883,14 @@
 #define tlb_unprotect_code tlb_unprotect_code_tricore
 #define tlb_reset_dirty tlb_reset_dirty_tricore
 #define tlb_set_dirty tlb_set_dirty_tricore
+#define tlb_set_page_full tlb_set_page_full_tricore
 #define tlb_set_page_with_attrs tlb_set_page_with_attrs_tricore
 #define tlb_set_page tlb_set_page_tricore
 #define get_page_addr_code_hostp get_page_addr_code_hostp_tricore
 #define get_page_addr_code get_page_addr_code_tricore
 #define probe_access probe_access_tricore
+#define probe_access_flags probe_access_flags_tricore
+#define probe_access_full probe_access_full_tricore
 #define tlb_vaddr_to_host tlb_vaddr_to_host_tricore
 #define tlb_vaddr_to_paddr tlb_vaddr_to_paddr_tricore
 #define helper_ret_ldub_mmu helper_ret_ldub_mmu_tricore
@@ -1354,14 +1357,23 @@
 #define helper_gvec_umax64 helper_gvec_umax64_tricore
 #define helper_gvec_bitsel helper_gvec_bitsel_tricore
 #define cpu_restore_state cpu_restore_state_tricore
+#define cpu_restore_state_from_tb cpu_restore_state_from_tb_tricore
+#define page_find_alloc page_find_alloc_tricore
+#define page_lock page_lock_tricore
+#define page_unlock page_unlock_tricore
+#define do_assert_page_locked do_assert_page_locked_tricore
+#define assert_no_pages_locked assert_no_pages_locked_tricore
 #define page_collection_lock page_collection_lock_tricore
 #define page_collection_unlock page_collection_unlock_tricore
 #define free_code_gen_buffer free_code_gen_buffer_tricore
 #define tcg_exec_init tcg_exec_init_tricore
+#define tb_htable_init tb_htable_init_tricore
 #define tb_cleanup tb_cleanup_tricore
 #define tb_flush tb_flush_tricore
 #define tb_phys_invalidate tb_phys_invalidate_tricore
 #define tb_gen_code tb_gen_code_tricore
+#define tb_reset_jump tb_reset_jump_tricore
+#define tb_link_page tb_link_page_tricore
 #define tb_exec_lock tb_exec_lock_tricore
 #define tb_exec_unlock tb_exec_unlock_tricore
 #define tb_invalidate_phys_page_range tb_invalidate_phys_page_range_tricore
