+++ symbols.sh
@@ -883,11 +883,14 @@ tlb_protect_code \
 tlb_unprotect_code \
 tlb_reset_dirty \
 tlb_set_dirty \
+tlb_set_page_full \
 tlb_set_page_with_attrs \
 tlb_set_page \
 get_page_addr_code_hostp \
 get_page_addr_code \
 probe_access \
+probe_access_flags \
+probe_access_full \
 tlb_vaddr_to_host \
 tlb_vaddr_to_paddr \
 helper_ret_ldub_mmu \
@@ -1354,14 +1357,23 @@ helper_gvec_umax32 \
 helper_gvec_umax64 \
 helper_gvec_bitsel \
 cpu_restore_state \
+cpu_restore_state_from_tb \
+page_find_alloc \
+page_lock \
+page_unlock \
+do_assert_page_locked \
+assert_no_pages_locked \
 page_collection_lock \
 page_collection_unlock \
 free_code_gen_buffer \
 tcg_exec_init \
+tb_htable_init \
 tb_cleanup \
 tb_flush \
 tb_phys_invalidate \
 tb_gen_code \
+tb_reset_jump \
+tb_link_page \
 tb_exec_lock \
 tb_exec_unlock \
 tb_invalidate_phys_page_range \
